feat: 리뷰 추가

+++ src/main/java/com/dev/kioki/domain/user/dto/UserResponseDTO.java
@@ -94,7 +94,7 @@ public class UserResponseDTO {
         String name;
         String imageUrl;
         Integer level;
-        Float rate;
+        Double rate;
         List<Long> stores_id;
     }
